+++ examples/lib/stories/input/joystick.dart
@@ -6,7 +6,7 @@ import 'package:flutter/painting.dart';
 
 import 'joystick_player.dart';
 
-class JoystickGame extends FlameGame with HasDraggableComponents {
+class JoystickGame extends FlameGame with HasDraggables {
   late final JoystickPlayer player;
   late final JoystickComponent joystick;